+++ VCU-APP/Core/Inc/Libs/_keyless.h
@@ -12,9 +12,9 @@
 #include "Drivers/_nrf24l01.h"
 
 /* Exported define -----------------------------------------------------------*/
-#define NRF_DATA_LENGTH											16U									// Max: 32 bytes
-#define NRF_ADDR_LENGTH                                          5U                                    // Range 3:5
-#define NRF_DATA_PAIR_LENGTH                                    (NRF_DATA_LENGTH + NRF_ADDR_LENGTH)
+#define NRF_DATA_LENGTH		    (uint8_t) 16									// Max: 32 bytes
+#define NRF_ADDR_LENGTH         (uint8_t) 5                                    // Range 3:5
+#define NRF_DATA_PAIR_LENGTH    (uint8_t) (NRF_DATA_LENGTH + NRF_ADDR_LENGTH)
 
 /* Exported enum -------------------------------------------------------------*/
 typedef enum {
@@ -31,12 +31,12 @@ typedef enum {
 typedef struct {
     nrf24l01_config config;
     struct {
-        uint8_t address[NRF_ADDR_LENGTH];
-        uint8_t payload[NRF_DATA_PAIR_LENGTH];
+        uint8_t address[NRF_ADDR_LENGTH ];
+        uint8_t payload[NRF_DATA_PAIR_LENGTH ];
     } tx;
     struct {
-        uint8_t address[NRF_ADDR_LENGTH];
-        uint8_t payload[NRF_DATA_PAIR_LENGTH];
+        uint8_t address[NRF_ADDR_LENGTH ];
+        uint8_t payload[NRF_DATA_PAIR_LENGTH ];
     } rx;
 } kless_t;
